separate host check and dns resolve
