+++ csm_test_utils/dns/__init__.py
@@ -1,0 +1,2 @@
+from .dns_resolving import main as dns_lookup_main
+from .host_check import main as host_main
